Expose game variable on resize
I believe the resize observer is triggered when the game is being loaded. This might be the perfect spot to expose game variables.
There might be more game variables for all I know but for my purposes this could possibly be all I need.

Everything we need can now be accessed through the global variable `exposedI` or `window.exposedI`

Hi, thanks for looking through the commit history. This project took place because of this video https://www.youtube.com/watch?v=gBHHSDJQNME?t=101 (specifically around 1:41 to 2:41)